Type Alias reth_db::abstraction::common::PairResult

pub type PairResult<T> = Result<Option<KeyValue<T>>, DatabaseError>;

A fallible key-value pair that may or may not exist.

The Result represents that the operation might fail, while the Option represents whether or not the entry exists.

Aliased Type

enum PairResult<T> {
    Ok(Option<(<T as Table>::Key, <T as Table>::Value)>),
    Err(DatabaseError),
}

Variants

Ok(Option<(<T as Table>::Key, <T as Table>::Value)>)

Contains the success value

Err(DatabaseError)

Contains the error value