Module reth_db::abstraction::cursor

Cursor database traits.

Structs

DupWalker

Provides an iterator to Cursor when handling a DupSort table.

RangeWalker

Provides a range iterator to Cursor when handling Table. Also check Walker

ReverseWalker

Provides a reverse iterator to Cursor when handling Table. Also check Walker

Walker

Provides an iterator to Cursor when handling Table.

Traits

DbCursorRO

A read-only cursor over table T.

DbCursorRW

Read write cursor over table.

DbDupCursorRO

A read-only cursor over the dup table T.

DbDupCursorRW

Read Write Cursor over DupSorted table.