Trait reth_db::abstraction::database_metrics::DatabaseMetrics

pub trait DatabaseMetrics {
    // Provided methods
    fn report_metrics(&self) { ... }
    fn gauge_metrics(&self) -> Vec<(&'static str, f64, Vec<Label>)> { ... }
    fn counter_metrics(&self) -> Vec<(&'static str, u64, Vec<Label>)> { ... }
    fn histogram_metrics(&self) -> Vec<(&'static str, f64, Vec<Label>)> { ... }
}

Represents a type that can report metrics, used mainly with the database. The report_metrics method can be used as a prometheus hook.

Provided Methods

fn report_metrics(&self)

Reports metrics for the database.

fn gauge_metrics(&self) -> Vec<(&'static str, f64, Vec<Label>)>

Returns a list of Gauge metrics for the database.

fn counter_metrics(&self) -> Vec<(&'static str, u64, Vec<Label>)>

Returns a list of Counter metrics for the database.

fn histogram_metrics(&self) -> Vec<(&'static str, f64, Vec<Label>)>

Returns a list of Histogram metrics for the database.