Trait reth_db::abstraction::table::Compress

pub trait Compress: Send + Sync + Sized + Debug {
    type Compressed: BufMut + AsRef<[u8]> + AsMut<[u8]> + Into<Vec<u8>> + Default + Send + Sync;

    // Required method
    fn compress_to_buf<B: BufMut + AsMut<[u8]>>(self, buf: &mut B);

    // Provided methods
    fn uncompressable_ref(&self) -> Option<&[u8]> { ... }
    fn compress(self) -> Self::Compressed { ... }
}

Trait that will transform the data to be saved in the DB in a (ideally) compressed format

Required Associated Types

type Compressed: BufMut + AsRef<[u8]> + AsMut<[u8]> + Into<Vec<u8>> + Default + Send + Sync

Compressed type.

Required Methods

fn compress_to_buf<B: BufMut + AsMut<[u8]>>(self, buf: &mut B)

Compresses data to a given buffer.

Provided Methods

fn uncompressable_ref(&self) -> Option<&[u8]>

If the type cannot be compressed, return its inner reference as Some(self.as_ref())

fn compress(self) -> Self::Compressed

Compresses data going into the database.

Object Safety

This trait is not object safe.