Trait reth_db::abstraction::table::DupSort

pub trait DupSort: Table {
    type SubKey: Key;
}

DupSort allows for keys to be repeated in the database.

Upstream docs: https://libmdbx.dqdkfa.ru/usage.html#autotoc_md48

Required Associated Types

type SubKey: Key

The table subkey. This type must implement Encode and Decode.

Sorting should be taken into account when encoding this.

Upstream docs: https://libmdbx.dqdkfa.ru/usage.html#autotoc_md48

Object Safety

This trait is not object safe.