Module reth_db::mdbx

Bindings for MDBX.

Modules

cursor

Cursor wrapper for libmdbx-sys.

ffi

tx

Transaction wrapper for libmdbx-sys.

Macros

mdbx_try_optional

Structs

CommitLatency

Commit latencies info.

Cursor

A cursor for navigating the items within a database.

Database

A handle to an individual database in an environment.

DatabaseEnv

Wrapper for the libmdbx environment: Environment

DatabaseFlags

Database options.

Environment

An environment supports multiple databases, all residing in the same shared-memory map.

EnvironmentBuilder

Options for opening or creating an environment.

EnvironmentFlags

Geometry

Info

Environment information.

ObjectLength

If you don’t need the data itself, just its length.

RO

RW

Stat

Environment statistics.

Transaction

An MDBX transaction.

WriteFlags

Write options.

Enums

DatabaseEnvKind

Environment used when opening a MDBX environment. RO/RW.

EnvironmentKind

Determines how data is mapped into memory

Error

An MDBX error kind.

HandleSlowReadersReturnCode

Iter

An iterator over the key/value pairs in an MDBX database.

IterDup

An iterator over the keys and duplicate values in an MDBX database.

Mode

PageSize

SyncMode

MDBX sync mode

Traits

TableObject

Implement this to be able to decode data values

TransactionKind

Type Aliases

HandleSlowReadersCallback

Handle-Slow-Readers callback function to resolve database full/overflow issue due to a reader(s) which prevents the old data from being recycled.

Result

An MDBX result.