Struct reth_db::mdbx::Cursor

pub struct Cursor<K>
where
    K: TransactionKind,
{ /* private fields */ }

A cursor for navigating the items within a database.

Implementations

impl<K> Cursor<K>
where K: TransactionKind,

pub fn cursor(&self) -> *mut MDBX_cursor

Returns a raw pointer to the underlying MDBX cursor.

The caller must ensure that the pointer is not used after the lifetime of the cursor.

pub fn iter_slices<'a>( &'a self ) -> IntoIter<'a, K, Cow<'a, [u8]>, Cow<'a, [u8]>>

Returns an iterator over the raw key value slices.

pub fn into_iter<Key, Value>(&self) -> IntoIter<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Returns an iterator over database items.

pub fn first<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at first key/data item.

pub fn first_dup<Value>(&mut self) -> Result<Option<Value>, Error>
where Value: TableObject,

DatabaseFlags::DUP_SORT-only: Position at first data item of current key.

pub fn get_both<Value>( &mut self, k: &[u8], v: &[u8] ) -> Result<Option<Value>, Error>
where Value: TableObject,

DatabaseFlags::DUP_SORT-only: Position at key/data pair.

pub fn get_both_range<Value>( &mut self, k: &[u8], v: &[u8] ) -> Result<Option<Value>, Error>
where Value: TableObject,

DatabaseFlags::DUP_SORT-only: Position at given key and at first data greater than or equal to specified data.

pub fn get_current<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Return key/data at current cursor position.

pub fn get_multiple<Value>(&mut self) -> Result<Option<Value>, Error>
where Value: TableObject,

DupFixed-only: Return up to a page of duplicate data items from current cursor position. Move cursor to prepare for Self::next_multiple().

pub fn last<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at last key/data item.

pub fn last_dup<Value>(&mut self) -> Result<Option<Value>, Error>
where Value: TableObject,

DupSort-only: Position at last data item of current key.

pub fn next<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at next data item

pub fn next_dup<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

DatabaseFlags::DUP_SORT-only: Position at next data item of current key.

pub fn next_multiple<Key, Value>( &mut self ) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

DatabaseFlags::DUP_FIXED-only: Return up to a page of duplicate data items from next cursor position. Move cursor to prepare for MDBX_NEXT_MULTIPLE.

pub fn next_nodup<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at first data item of next key.

pub fn prev<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at previous data item.

pub fn prev_dup<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

DatabaseFlags::DUP_SORT-only: Position at previous data item of current key.

pub fn prev_nodup<Key, Value>(&mut self) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at last data item of previous key.

pub fn set<Value>(&mut self, key: &[u8]) -> Result<Option<Value>, Error>
where Value: TableObject,

Position at specified key.

pub fn set_key<Key, Value>( &mut self, key: &[u8] ) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at specified key, return both key and data.

pub fn set_range<Key, Value>( &mut self, key: &[u8] ) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at first key greater than or equal to specified key.

pub fn prev_multiple<Key, Value>( &mut self ) -> Result<Option<(Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

DatabaseFlags::DUP_FIXED-only: Position at previous page and return up to a page of duplicate data items.

pub fn set_lowerbound<Key, Value>( &mut self, key: &[u8] ) -> Result<Option<(bool, Key, Value)>, Error>
where Key: TableObject, Value: TableObject,

Position at first key-value pair greater than or equal to specified, return both key and data, and the return code depends on a exact match.

For non DupSort-ed collections this works the same as Self::set_range(), but returns false if key found exactly and true if greater key was found.

For DupSort-ed a data value is taken into account for duplicates, i.e. for a pairs/tuples of a key and an each data value of duplicates. Returns false if key-value pair found exactly and true if the next pair was returned.

pub fn iter<Key, Value>(&mut self) -> Iter<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Returns an iterator over database items.

The iterator will begin with item next after the cursor, and continue until the end of the database. For new cursors, the iterator will begin with the first item in the database.

For databases with duplicate data items (DatabaseFlags::DUP_SORT), the duplicate data items of each key will be returned before moving on to the next key.

pub fn iter_start<Key, Value>(&mut self) -> Iter<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Iterate over database items starting from the beginning of the database.

For databases with duplicate data items (DatabaseFlags::DUP_SORT), the duplicate data items of each key will be returned before moving on to the next key.

pub fn iter_from<Key, Value>(&mut self, key: &[u8]) -> Iter<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Iterate over database items starting from the given key.

For databases with duplicate data items (DatabaseFlags::DUP_SORT), the duplicate data items of each key will be returned before moving on to the next key.

pub fn iter_dup<Key, Value>(&mut self) -> IterDup<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Iterate over duplicate database items. The iterator will begin with the item next after the cursor, and continue until the end of the database. Each item will be returned as an iterator of its duplicates.

pub fn iter_dup_start<Key, Value>(&mut self) -> IterDup<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Iterate over duplicate database items starting from the beginning of the database. Each item will be returned as an iterator of its duplicates.

pub fn iter_dup_from<Key, Value>( &mut self, key: &[u8] ) -> IterDup<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Iterate over duplicate items in the database starting from the given key. Each item will be returned as an iterator of its duplicates.

pub fn iter_dup_of<Key, Value>(&mut self, key: &[u8]) -> Iter<'_, K, Key, Value>
where Key: TableObject, Value: TableObject,

Iterate over the duplicates of the item in the database with the given key.

impl Cursor<RW>

pub fn put( &mut self, key: &[u8], data: &[u8], flags: WriteFlags ) -> Result<(), Error>

Puts a key/data pair into the database. The cursor will be positioned at the new data item, or on failure usually near it.

pub fn del(&mut self, flags: WriteFlags) -> Result<(), Error>

Deletes the current key/data pair.

Flags

WriteFlags::NO_DUP_DATA may be used to delete all data items for the current key, if the database was opened with DatabaseFlags::DUP_SORT.