Module reth_db::snapshot

reth’s snapshot database table import and access

Macros

add_snapshot_mask

Add mask to select N column values from a specific snapshot segment row.

Structs

HeaderMask

Mask for Header snapshot segment. See Mask for more.

Mask

Generic Mask helper struct for selecting specific column values to read and decompress.

ReceiptMask

Mask for Receipt snapshot segment. See Mask for more.

SnapshotCursor

Cursor of a snapshot segment.

TransactionMask

Mask for Transaction snapshot segment. See Mask for more.

Traits

ColumnSelectorOne

Trait for specifying a mask to select one column value.

ColumnSelectorThree

Trait for specifying a mask to select three column values.

ColumnSelectorTwo

Trait for specifying a mask to select two column values.

Functions

create_snapshot_T1

Creates a snapshot from specified tables. Each table’s Value iterator represents a column.

create_snapshot_T1_T2

Creates a snapshot from specified tables. Each table’s Value iterator represents a column.

create_snapshot_T1_T2_T3

Creates a snapshot from specified tables. Each table’s Value iterator represents a column.

create_snapshot_T1_T2_T3_T4

Creates a snapshot from specified tables. Each table’s Value iterator represents a column.

create_snapshot_T1_T2_T3_T4_T5

Creates a snapshot from specified tables. Each table’s Value iterator represents a column.

iter_snapshots

Given the snapshots directory path, it returns a list over the existing snapshots organized by SnapshotSegment. Each segment has a sorted list of block ranges and transaction ranges.