Function reth_db::snapshot::create_snapshot_T1_T2_T3_T4_T5

pub fn create_snapshot_T1_T2_T3_T4_T5<T1: Table<Key = K>, T2: Table<Key = K>, T3: Table<Key = K>, T4: Table<Key = K>, T5: Table<Key = K>, K, H: NippyJarHeader>(
    tx: &impl DbTx,
    range: RangeInclusive<K>,
    additional: Option<Vec<Box<dyn Iterator<Item = Result<Vec<u8>, Box<dyn StdError + Send + Sync>>>>>>,
    dict_compression_set: Option<Vec<impl Iterator<Item = Vec<u8>>>>,
    keys: Option<impl Iterator<Item = ColumnResult<impl PHFKey>>>,
    row_count: usize,
    nippy_jar: &mut NippyJar<H>
) -> ProviderResult<()>
where
    K: Key + Copy,

Creates a snapshot from specified tables. Each table’s Value iterator represents a column.

Ensure the range contains the same number of rows.

tx: Database transaction.

range: Data range for columns in tables.

additional: Additional columns which can’t be straight straightforwardly walked on.

keys: Iterator of keys (eg. TxHash or BlockHash) with length equal to row_count and ordered by future column insertion from range.

dict_compression_set: Sets of column data for compression dictionaries. Max size is 2GB. Row count is independent.

row_count: Total rows to add to NippyJar. Must match row count in range.

nippy_jar: Snapshot object responsible for file generation.