Trait reth_db::snapshot::ColumnSelectorOne

pub trait ColumnSelectorOne {
    type FIRST: Decompress;

    const MASK: usize;
}

Trait for specifying a mask to select one column value.

Required Associated Types

type FIRST: Decompress

First desired column value

Required Associated Constants

const MASK: usize

Mask to obtain desired values, should correspond to the order of columns in a snapshot.

Object Safety

This trait is not object safe.