Trait reth_db::snapshot::ColumnSelectorThree

pub trait ColumnSelectorThree {
    type FIRST: Decompress;
    type SECOND: Decompress;
    type THIRD: Decompress;

    const MASK: usize;
}

Trait for specifying a mask to select three column values.

Required Associated Types

type FIRST: Decompress

First desired column value

type SECOND: Decompress

Second desired column value

type THIRD: Decompress

Third desired column value

Required Associated Constants

const MASK: usize

Mask to obtain desired values, should correspond to the order of columns in a snapshot.

Object Safety

This trait is not object safe.