Trait reth_db::snapshot::ColumnSelectorTwo

pub trait ColumnSelectorTwo {
    type FIRST: Decompress;
    type SECOND: Decompress;

    const MASK: usize;
}

Trait for specifying a mask to select two column values.

Required Associated Types

type FIRST: Decompress

First desired column value

type SECOND: Decompress

Second desired column value

Required Associated Constants

const MASK: usize

Mask to obtain desired values, should correspond to the order of columns in a snapshot.

Object Safety

This trait is not object safe.