Function reth_db::version::create_db_version_file

pub fn create_db_version_file<P: AsRef<Path>>(db_path: P) -> Result<()>

Creates a database version file with DB_VERSION_FILE_NAME name containing DB_VERSION string.

This function will create a file if it does not exist, and will entirely replace its contents if it does.